Updates to modal component
